checkout method is added

+++ Patterns/src/Payment.java
@@ -1,3 +1,3 @@
 public interface Payment {
-    public void pay(int amount);
+    public void pay(double amount);
 }
