+++ Patterns/src/RoachPal.java
@@ -13,7 +13,7 @@ public class RoachPal implements Payment{
     }
 
     @Override
-    public void pay(int amount) {
+    public void pay(double amount) {
         System.out.println("Payment from RoachPal" + amount);
     }
 }
